bibliothèque_en_ligne/projectbib/biblio/src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==0.5.0
Django==4.2.15
django-celery-results==2.5.1
django-configurations==2.5.1
django-cors-headers==3.11.0
django-filter==21.1
django-model-utils==4.2.0
django-rest-passwordreset==1.2.1
django-storages==1.12.3
django-unique-upload==0.2.1
djangorestframework==3.13.1
kombu==5.4.0
msgpack==1.0.8
prompt-toolkit==3.0.47
psycopg2-binary==2.9.3
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
